requirements.txt
requests
configobj
ruamel.yaml